RAG_o1/requirements.txt
# Core RAG and Vector DB
langchain>=0.1.0
langchain-community>=0.0.10
qdrant-client>=1.6.0
sentence-transformers>=2.2.2

# Document Processing
python-dotenv>=1.0.0
beautifulsoup4>=4.12.2
markdown>=3.5.1
pdfminer.six>=20221105

# Error Handling and Logging
loguru>=0.7.2

# Data Processing
pandas>=2.1.3
numpy>=1.24.3

# ML and Embeddings
torch>=2.1.1
transformers>=4.35.2

# Visualization
networkx>=3.2.1
matplotlib>=3.8.2

# Type Checking
pydantic>=2.5.2
typing-extensions>=4.8.0

# Testing
pytest>=7.4.3
pytest-cov>=4.1.0
